refactor: move slices to special directory
