+++ src/pages/favorites-page/cards/favorite-offer-card.tsx
@@ -1,7 +1,7 @@
 import {FC} from 'react';
 import {Link} from 'react-router-dom';
 import {Offer} from '../../../entities/offer/model/types.ts';
-import {changeFavoriteStatus} from '../../../features/offers-manager/model/favorites-page-slice.ts';
+import {changeFavoriteStatus} from '../../../slices/favorites-page-slice.ts';
 import {RoutePath} from '../../../shared/enums/routes.ts';
 import {useAppDispatch} from '../../../shared/redux-helpers/typed-hooks.ts';
 import {FavoriteStatus} from '../../../shared/server-interaction/constants.ts';
