+++ src/pages/main-page/places-container.tsx
@@ -4,8 +4,8 @@ import {DEFAULT_CITY} from '../../entities/city/model/constants.ts';
 import {OfferSortOption} from '../../entities/offer/model/constants.ts';
 import {Offer} from '../../entities/offer/model/types.ts';
 import {OfferCardList} from '../../entities/offer/ui/offer-card-list.tsx';
-import {changeFavoriteStatus} from '../../features/offers-manager/model/favorites-page-slice.ts';
-import {setActiveOffer, setOffersSort} from '../../features/offers-manager/model/offers-slice.ts';
+import {changeFavoriteStatus} from '../../slices/favorites-page-slice.ts';
+import {setActiveOffer, setOffersSort} from '../../slices/offers-slice.ts';
 import {RoutePath} from '../../shared/enums/routes.ts';
 import {useAppDispatch, useAppSelector} from '../../shared/redux-helpers/typed-hooks.ts';
 import {FavoriteStatus} from '../../shared/server-interaction/constants.ts';
