+++ src/pages/offer-page/comments-list.tsx
@@ -1,7 +1,7 @@
 import {useEffect} from 'react';
 import {useParams} from 'react-router-dom';
 import {CommentaryCard} from '../../entities/commentary/ui/commentary-card.tsx';
-import {loadComments} from '../../features/offers-manager/model/offer-page-slice.ts';
+import {loadComments} from '../../slices/offer-page-slice.ts';
 import {useAppDispatch, useAppSelector} from '../../shared/redux-helpers/typed-hooks.ts';
 import {Spinner} from '../../widgets/spinner/ui/spinner.tsx';
