+++ src/pages/offer-page/offer-page.tsx
@@ -2,8 +2,8 @@ import {FC, useCallback, useEffect} from 'react';
 import {Navigate, useNavigate, useParams} from 'react-router-dom';
 import {Offer, OfferExtendedDto} from '../../entities/offer/model/types.ts';
 import {OfferCardList} from '../../entities/offer/ui/offer-card-list.tsx';
-import {changeFavoriteStatus} from '../../features/offers-manager/model/favorites-page-slice.ts';
-import {loadNearbyOffers, loadOffer} from '../../features/offers-manager/model/offer-page-slice.ts';
+import {changeFavoriteStatus} from '../../slices/favorites-page-slice.ts';
+import {loadNearbyOffers, loadOffer} from '../../slices/offer-page-slice.ts';
 import {RoutePath} from '../../shared/enums/routes.ts';
 import {useAppDispatch, useAppSelector} from '../../shared/redux-helpers/typed-hooks.ts';
 import {FavoriteStatus} from '../../shared/server-interaction/constants.ts';
